move halo2 crate to a backend module
