+++ Cargo.toml
@@ -15,7 +15,6 @@ members = [
     "pipeline",
     "pilopt",
     "asm-to-pil",
-    "halo2",
     "backend",
     "ast",
     "analysis",
