+++ backend/Cargo.toml
@@ -8,15 +8,27 @@ homepage = { workspace = true }
 repository = { workspace = true }
 
 [features]
-halo2 = ["dep:powdr-halo2"]
+default = []
+halo2 = ["dep:halo2_proofs", "dep:halo2_curves", "dep:snark-verifier", "dep:halo2_solidity_verifier"]
 estark-polygon = ["dep:pil-stark-prover"]
 
 [dependencies]
 powdr-ast = { path = "../ast" }
-powdr-halo2 = { path = "../halo2", optional = true }
 powdr-number = { path = "../number" }
 powdr-pil-analyzer = { path = "../pil-analyzer" }
 powdr-executor = { path = "../executor" }
+
+starky = { git = "https://github.com/0xEigenLabs/eigen-zkvm.git", rev = "59d2152" }
+pil-stark-prover = { git = "https://github.com/powdr-labs/pil-stark-prover.git", rev = "769b1153f3ae2d7cbab4c8acf33865ed13f8a823", optional = true }
+
+# TODO change this once Halo2 releases 0.3.1
+#halo2_proofs = { git = "https://github.com/privacy-scaling-explorations/halo2", tag = "v0.3.0", features = ["circuit-params"] }
+halo2_proofs = { git = "https://github.com/powdr-labs/halo2", branch = "make-emit-public-v0.3.0", features = ["circuit-params"], optional = true }
+halo2_curves = { version = "0.6.1", package = "halo2curves", optional = true}
+# TODO change this once Halo2 releases 0.3.1 and snark-verifier uses it
+#snark-verifier = { git = "https://github.com/privacy-scaling-explorations/snark-verifier", tag = "v2024_01_31" }
+snark-verifier = { git = "https://github.com/powdr-labs/snark-verifier", branch = "powdr-halo2", optional = true }
+halo2_solidity_verifier = { git = "https://github.com/powdr-labs/halo2-solidity-verifier", branch = "main", features = ["evm"], optional = true }
 
 strum = { version = "0.24.1", features = ["derive"] }
 log = "0.4.17"
@@ -24,9 +36,11 @@ serde = "1.0"
 serde_json = "1.0"
 hex = "0.4"
 thiserror = "1.0.43"
-starky = { git = "https://github.com/0xEigenLabs/eigen-zkvm.git", rev = "59d2152" }
-pil-stark-prover = { git = "https://github.com/powdr-labs/pil-stark-prover.git", rev = "769b1153f3ae2d7cbab4c8acf33865ed13f8a823", optional = true }
 mktemp = "0.5.0"
+num-traits = "0.2.15"
+num-integer = "0.1.45"
+itertools = "^0.10"
+rand = "0.8.5"
 
 [dev-dependencies]
 test-log = "0.2.12"
