--- halo2/Cargo.toml
@@ -1,45 +1,0 @@
-[package]
-name = "powdr-halo2"
-description = "powdr Halo2 integration"
-version = { workspace = true }
-edition = { workspace = true }
-license = { workspace = true }
-homepage = { workspace = true }
-repository = { workspace = true }
-
-[dependencies]
-powdr-ast = { path = "../ast" }
-powdr-number = { path = "../number" }
-powdr-executor = { path = "../executor" }
-
-# TODO change this once Halo2 releases 0.3.1
-#halo2_proofs = { git = "https://github.com/privacy-scaling-explorations/halo2", tag = "v0.3.0", features = ["circuit-params"] }
-halo2_proofs = { git = "https://github.com/powdr-labs/halo2", branch = "make-emit-public-v0.3.0", features = ["circuit-params"] }
-halo2_curves = { version = "0.6.1", package = "halo2curves" }
-# TODO change this once Halo2 releases 0.3.1 and snark-verifier uses it
-#snark-verifier = { git = "https://github.com/privacy-scaling-explorations/snark-verifier", tag = "v2024_01_31" }
-snark-verifier = { git = "https://github.com/powdr-labs/snark-verifier", branch = "powdr-halo2" }
-halo2_solidity_verifier = { git = "https://github.com/powdr-labs/halo2-solidity-verifier", branch = "main", features = ["evm"] }
-
-num-traits = "0.2.15"
-num-integer = "0.1.45"
-itertools = "^0.10"
-log = "0.4.17"
-rand = "0.8.5"
-serde = "1.0"
-serde_json = "1.0"
-
-[dev-dependencies]
-powdr-airgen = { path = "../airgen" }
-powdr-analysis = { path = "../analysis" }
-powdr-executor = { path = "../executor" }
-powdr-importer = { path = "../importer" }
-powdr-linker = { path = "../linker" }
-powdr-parser = { path = "../parser" }
-powdr-pipeline = { path = "../pipeline" }
-
-test-log = "0.2.12"
-env_logger = "0.10.0"
-
-[package.metadata.cargo-udeps.ignore]
-development = ["env_logger"]
