--- halo2/src/lib.rs
@@ -1,9 +1,0 @@
-#![deny(clippy::print_stdout)]
-
-pub(crate) mod aggregation;
-pub(crate) mod circuit_builder;
-pub(crate) mod mock_prover;
-pub(crate) mod prover;
-
-pub use mock_prover::mock_prove;
-pub use prover::*;
